project-root/fetcher/requirements.txt
pandas
requests
sqlalchemy
pymysql
python-dotenv
cryptography